add remote
Refactor getArticle and createArticle functions, update article image
properties

This commit refactors the `getArticle` and `createArticle` functions in
the `app/models/article.server.ts` file. The functions now use the
updated property names `image_sm` and `image_lg` instead of the
deprecated `articleImage`. This change improves clarity and consistency
in the codebase. The update is necessary to align with the changes made
to the data structure in the `prisma/data.ts` file.

No issue references.